Resolved Conflicts in the following files:
	alljoyn_core/router/AllJoynObj.cc
	alljoyn_core/router/AllJoynObj.h
	alljoyn_core/router/ns/IpNameServiceImpl.cc
	alljoyn_core/src/AllJoynStd.cc
	services/about/cpp/src/AnnouncementRegistrar.cc

Updated alljoyn_core/router/SessionlessObj.cc due to a build failure
that resulted from the merge.

+++ services/about/java/samples/java/AboutServerSample/src/org/alljoyn/about/sample/AboutServerSample.java
@@ -1,5 +1,5 @@
 /******************************************************************************
- * Copyright (c) 2013, AllSeen Alliance. All rights reserved.
+ * Copyright (c) 2013-2014, AllSeen Alliance. All rights reserved.
  *
  *    Permission to use, copy, modify, and/or distribute this software for any
  *    purpose with or without fee is hereby granted, provided that the above
@@ -357,7 +357,7 @@ public class AboutServerSample {
 	         
 	      
 	         try{
-	        	 AboutServiceImpl.getInstance().startAboutServer((short)CONTACT_PORT, impl, mBus);	 
+	        	 AboutServiceImpl.getInstance().startAboutServer(CONTACT_PORT, impl, mBus);	 
 	         }catch (Exception e){
 	        	 e.printStackTrace();
 	         }
